# Optimize asHexString method