The method asHexString has been rewrite. `String.format` has been removed and the method was now around 60 times faster.


Benchmarked with a 100K iteration with the following code:
```java
for (int j = 0; j < 3; ++j) {
    startTime = System.currentTimeMillis();
    for (int i = 0; i < 100000; ++i) {
        asHexString_OLD(data);
    }
    endTime = System.currentTimeMillis();
    LOG.info("asHexString_OLD -> {}ms", endTime - startTime);
}
for (int j = 0; j < 3; ++j) {
    startTime = System.currentTimeMillis();
    for (int i = 0; i < 100000; ++i) {
        asHexString_NEW(data);
    }
    endTime = System.currentTimeMillis();
    LOG.info("asHexString_NEW -> {}ms", endTime - startTime);
}
```

```console
[2017-04-22 21:04:42] INFO - MainEntry - asHexString_OLD -> 3130ms
[2017-04-22 21:04:44] INFO - MainEntry - asHexString_OLD -> 2615ms
[2017-04-22 21:04:47] INFO - MainEntry - asHexString_OLD -> 2573ms
[2017-04-22 21:04:47] INFO - MainEntry - asHexString_NEW -> 40ms
[2017-04-22 21:04:47] INFO - MainEntry - asHexString_NEW -> 24ms
[2017-04-22 21:04:47] INFO - MainEntry - asHexString_NEW -> 22ms
```